WARNING (MIGHT NOT WORK): This commit could not work properly. The tests work with a locally running docker image (see provided docker-compose.yml file) but the CI pipeline could be configured wrongly.
CHANGES:
This refactors the db handler tests to use fixtures for the db setup
and adds the migration test for the postgresql handler. This tests
builds upon a running postgresql instance, therefore a docker-compose
file is provided to test everything locally without CI pipeline.

The commit is rather large, note to myself: make smaller commits even
if that breaks the running code sometimes.