TESTING: there seem to be rare edge cases (appears randomly) where postgres tests fail because of some database setup issues.
Refactor test_monitor_handler.py: Refactor reset_db() (db_cleanup)
function. In some cases there are issues, when running tests where
the public schema cannot be found anymore or seems not the be set
as default schema. THIS PROBLEM STILL OCCURES SOMETIMES. I was not able
to reproduce it / find the cause appropriately. I hope it is a local
db error but in case there are bugs concerning this in the future,
this function could be a place to start at.