Fix: PostgresDBHandler.get_env_id() was returning Tuple where id asked.
The function gets a tuple from the database query and thus
needs to unpack the tuple and only return the inside.

It would be far much easier if Python just enforced typing on function
in and out parameters, but I guess that's a little too late now...
maybe for Python 4.