Fix: Fix 3 issues: with-context, db-close and env-id problem
with-context: Remove the with-contexts inside the PostgresDBHandler and
do manual commit after cursor execution instead to prevent closing of
the database connection.

db-close: Add close() function to both db-handlers and the
PytestMonitorSession class in order to properly close all database
connections at the end of the tests. This happens in pytest-monitor.py
with the pytest_sessionfinish hook implementation that calls the
close() function on the PytestMonitorSession object.

env-id: The query for the env-id PostgresDBHandler.get_env_id() queries
for the env id in the database and gets a tuple from the query but does
return it without unwrapping the value inside. This leads to problems
as the value is meant to be unwrapped, therefore return the unwrapped
value.