Add tests for session and database implementations (dbhandlers)
The session is tested for proper closing of the database connection.
The close() function is called from the hookimplementation in
pytest_sessionfinish().

The dbhandlers are tested for proper setup of the tables and in
consequence a working connection. The PostgresDBHandler requires a
running postgres server, that's what the docker-compose.yml file is for