Add workaround fix for faulty memory_profiler module
The memory profiler only reports Exceptions when it should report all
Exceptions that inherit from BaseExceptions.

This is ultimately reworked in a PR incorporating a simplified
memory profiler module into the codebase that fixes not only this issue
but also gives the possibility of getting measurements for failing
tests.

This workaround uses return values to work around the issue. This way
pytest-monitor can check for BaseExceptions and act accordingly.
Like described earlier the ultimate goal should probably be to replace
the whole memory profiler as proposed in this PR:
https://github.com/CFMTech/pytest-monitor/pull/82